% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INV_get_id.R
\name{INV_get_id}
\alias{INV_get_id}
\title{Retrieve the Investing.com ID for a Financial Asset}
\usage{
INV_get_id(info, language = "en", exchange = NULL)
}
\arguments{
\item{info}{Character. The search term or financial asset (e.g., stock symbol) to query. Must be a string.}

\item{language}{Character. The language for the search results. Defaults to "en" for English. Other languages can be passed (e.g., "fr" for French).}

\item{exchange}{Character. Optional. The stock exchange to filter results by. If specified, the results will be filtered to match the provided exchange name.}
}
\value{
A character vector representing the ID of the financial asset from Investing.com. If the asset does not exist, an error is thrown.
}
\description{
This function retrieves the Investing.com ID of a specified financial asset based on the provided search term and optional exchange.
}
\examples{
library(httr)
library(jsonlite)
library(rvest)
library(glue)
library(magrittr)

# Retrieve the ID of "aaple" (Apple Inc.)
INV_get_id("aaple")

# Retrieve the ID of 'SNTS' from BRVM exchange
INV_get_id("SNTS", exchange = "BRVM")

}
\seealso{
\code{\link{INV_get_info}} for retrieving more detailed information about the asset.
}
\author{
Koffi Frederic SESSIE
}
