% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INV_get_info.R
\name{INV_get_info}
\alias{INV_get_info}
\title{Retrieve Information from Investing.com}
\usage{
INV_get_info(info, language = "en", exchange = NULL)
}
\arguments{
\item{info}{Character. The search term or financial asset (e.g., stock symbol, ISIN) to query. Must be a string.}

\item{language}{Character. The language for the search results. Defaults to "en" for English. Other languages can be passed (e.g., "fr" for French).}

\item{exchange}{Character. Optional. The stock exchange to filter results by. If specified, the results will be filtered to match the provided exchange name.}
}
\value{
A data frame containing financial information related to the search query, including:
\item{id}{The identifier of the asset on Investing.com.}
\item{url}{The URL to the asset's page on Investing.com.}
\item{description}{A brief description of the asset.}
\item{symbol}{The asset's ticker symbol.}
\item{exchange}{The stock exchange where the asset is traded.}
\item{flag}{The country or region of the asset.}
\item{type}{The type of financial asset (e.g., stock).}
\item{isin}{The asset's ISIN (International Securities Identification Number), if available.}
\item{ytd}{The asset's Year-To-Date (YTD) performance, if available.}
}
\description{
This function retrieves financial information, including ISIN and Year-To-Date (YTD) performance data, from Investing.com based on a search query.
}
\examples{
library(httr)
library(jsonlite)
library(rvest)
library(glue)
library(magrittr)

# Retrieve information on Ecobank Transnational Inc.
INV_get_info("ETIT")

# Retrieve information on a financial asset in French
INV_get_info("ETIT", language = "fr")

\donttest{
# Retrieve 'societe general sa' information on EuroTLX exchange
INV_get_info("societe general sa", exchange = "EuroTLX")
}

}
\author{
Koffi Frederic SESSIE
}
