% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INV_hcDt.R
\name{INV_hcDt}
\alias{INV_hcDt}
\title{Fetch a Ticker's Historical Data from Investing.com}
\usage{
INV_hcDt(ticker_id, interval = NULL, period = NULL, max_retries = 5)
}
\arguments{
\item{ticker_id}{The ticker identifier as a number. This is the unique ID for the financial asset on Investing.com.}

\item{interval}{A string representing the interval between data points.
It can be one of the following values:
\itemize{
\item "PT1M" (1 minute),
\item "PT5M" (5 minutes),
\item "PT15M" (15 minutes),
\item "PT30M" (30 minutes),
\item "PT1H" (1 hour),
\item "PT5H" (5 hours),
\item "P1D" (1 day),
\item "P1W" (1 week),
\item "P1M" (1 month).
}
Defaults to \code{NULL}. If \code{period} is set, the function automatically adjusts \code{interval}.}

\item{period}{A string representing the total time range for the data.
It can be one of the following values:
\itemize{
\item "P1Y" (1 year),
\item "P5Y" (5 years),
\item "MAX" (maximum available data),
\item "P1D" (1 day),
\item "P1W" (1 week),
\item "P1M" (1 month).
}
Defaults to \code{NULL}. If provided, the function adjusts \code{interval} accordingly.}

\item{max_retries}{Integer value representing the maximum number of retries in case of a failed API request. Default is 5.}
}
\value{
A tibble with the following columns:
\itemize{
\item \code{Date}: The date and time of the data point.
\item \code{Open}: The opening price.
\item \code{High}: The highest price during the interval.
\item \code{Low}: The lowest price during the interval.
\item \code{Close}: The closing price.
\item \code{Volume}: The trading volume during the interval.
\item \code{Extra_info}: Optional additional information (only included if present in the API response).
}
}
\description{
This function retrieves historical data for a given financial ticker from Investing.com using their public API.
It allows specifying an interval and period to adjust the frequency of the data points.
}
\details{
The function queries Investing.com's API for historical financial data based on the given parameters.
It allows for the retrieval of data at various intervals and over different periods.
The function supports retry attempts in case of a failed API call.
}
\examples{
library(httr)
library(jsonlite)
library(tibble)
library(glue)
library(rlang)
#\donttest{
# Fetch historical data daily for France 5Y OAT
data <- INV_hcDt(ticker_id = 23769) #23763 is OAT 5y id
head(data)

# Fetch historical data for France 5Y OAT over 1 year with weekly interval
data_w <- INV_hcDt(ticker_id = 23769, interval <- 'P1W', period = "P1Y")
head(data_w)
#}

}
\author{
Koffi Frederic SESSIE
}
